mc: Allow setting the TET_ROOT at build time
meson doesn't allow us to pass an environment variable, so we need
another method. I tried using a python wrapper, but that doesn't work
with a generator because there's no way to correctly describe the
dependency. Instead I've justed added a #define that we can se in
meson to get the behavior we want.